requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
requests==2.31.0
pyyaml==6.0.1
python-dateutil==2.8.2
websockets==12.0

# Database
sqlalchemy==2.0.23

# CTI & Threat Intelligence
python-whois==0.8.0
dnspython==2.4.2
geoip2==4.7.0

# MITRE ATT&CK
stix2==3.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging
colorlog==6.8.0

# PDF Report Generation
reportlab==4.0.7
Pillow==10.1.0

# Windows Event Log Collection (Windows only)
pywin32==306; sys_platform == 'win32'
psutil==5.9.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
